Refactored name of anim and anim panel to Menu. Removed unused class leaderboard.java. Added javadocs for public methods in aforementioned menu, anim panel and loginwindow classes.
