--- src/main/java/views/Leaderboard.java
@@ -1,4 +1,0 @@
-package menu;
-
-public class Leaderboard {
-}
